refactor(components): keep collection search normalization private
Rename module to matchesCollectionSearch and test only the exported matcher.
