+++ packages/components/src/templates/next/layouts/Collection/utils/__tests__/matchesCollectionSearch.test.ts
@@ -1,79 +1,22 @@
 import { describe, expect, it } from "vitest"
 
-import {
-  matchesCollectionSearch,
-  normalizeCollectionSearchText,
-} from "../normalizeCollectionSearchText"
-
-describe("normalizeCollectionSearchText", () => {
-  it("lowercases and trims text", () => {
-    // Arrange
-    const text = "  Guide To Isomer  "
-
-    // Act
-    const result = normalizeCollectionSearchText(text)
-
-    // Assert
-    expect(result).toBe("guide to isomer")
-  })
-
-  it("normalizes fullwidth parentheses to ASCII", () => {
-    // Arrange
-    const text = "MANAGEMENT（FM）"
-
-    // Act
-    const result = normalizeCollectionSearchText(text)
-
-    // Assert
-    expect(result).toBe("management(fm)")
-  })
-
-  it("normalizes a non-breaking space before parentheses", () => {
-    // Arrange
-    const text = "MANAGEMENT\u00a0(FM)"
-
-    // Act
-    const result = normalizeCollectionSearchText(text)
-
-    // Assert
-    expect(result).toBe("management(fm)")
-  })
-
-  it("removes soft hyphens", () => {
-    // Arrange
-    const text = "MANAGEMENT\u00ad(FM)"
-
-    // Act
-    const result = normalizeCollectionSearchText(text)
-
-    // Assert
-    expect(result).toBe("management(fm)")
-  })
-
-  it("removes zero-width spaces", () => {
-    // Arrange
-    const text = "MANAGEMENT\u200b(FM)"
-
-    // Act
-    const result = normalizeCollectionSearchText(text)
-
-    // Assert
-    expect(result).toBe("management(fm)")
-  })
-
-  it("treats missing space before parentheses as equivalent to spaced text", () => {
-    // Arrange
-    const withSpace = normalizeCollectionSearchText("MANAGEMENT (FM)")
-    const withoutSpace = normalizeCollectionSearchText("MANAGEMENT(FM)")
-
-    // Assert
-    expect(withSpace).toBe(withoutSpace)
-  })
-})
+import { matchesCollectionSearch } from "../matchesCollectionSearch"
 
 describe("matchesCollectionSearch", () => {
   const circularTitle =
     "CIRCULAR ON NEW FEEDBACK CHANNEL ON PUBLIC SECTOR FACILITIES MANAGEMENT (FM) PROJECTS AND REVISED GUIDE ON FM PROCUREMENT"
+
+  it("matches case-insensitively and ignores surrounding whitespace in the query", () => {
+    // Arrange
+    const title = "Guide to Isomer"
+    const search = "  guide to isomer  "
+
+    // Act
+    const result = matchesCollectionSearch(title, search)
+
+    // Assert
+    expect(result).toBe(true)
+  })
 
   it("matches a full title search that includes parentheses", () => {
     // Arrange
@@ -131,6 +74,30 @@ describe("matchesCollectionSearch", () => {
     const title = circularTitle.replace(" (FM)", "(FM)")
     const search =
       "CIRCULAR ON NEW FEEDBACK CHANNEL ON PUBLIC SECTOR FACILITIES MANAGEMENT (FM)"
+
+    // Act
+    const result = matchesCollectionSearch(title, search)
+
+    // Assert
+    expect(result).toBe(true)
+  })
+
+  it("matches when the stored title contains a soft hyphen before parentheses", () => {
+    // Arrange
+    const title = "MANAGEMENT\u00ad(FM) PROJECTS"
+    const search = "management (FM)"
+
+    // Act
+    const result = matchesCollectionSearch(title, search)
+
+    // Assert
+    expect(result).toBe(true)
+  })
+
+  it("matches when the stored title contains a zero-width space before parentheses", () => {
+    // Arrange
+    const title = "MANAGEMENT\u200b(FM) PROJECTS"
+    const search = "management (FM)"
 
     // Act
     const result = matchesCollectionSearch(title, search)
